feat: check/create config file in default location
If a config file is not passed explicitly, ~/.config/ochami/ is checked
for config.<format> where <format> is a file format passed via
--config-format. This is any format accepted by viper (e.g. yaml, json,
ini, etc.) and is yaml by default.

If the config file, whether specified via the command line or set as the
default in ~/.config/ochami, does not exist, the user is asked if they
want it to be created and populated with the default values.