fix(discover): explicitly add components before redfish endpoints
SMD will create the components based on info in the redfish endpoints.
However, redfish endpoints do not contain NID information and so SMD
creates the components with NIDs starting at 1 and incrememnting by 1
with each component added. If the NID is specified in the discovery
file, we do not want this behavior. Thus, this commit explicitly adds
the components with the specified NID *before* adding the redfish
endpoints.