feat: group SMD and BSS commands
Group "low-level" commands, commands that send a single request to
SMD/BSS and return the result directly, under the respective service
metacommand. Commands that send a request to BSS are prefixed with 'bss'
while commands that send a request to SMD are prefixed with 'smd'.

This will save room for the more "high level" commands, commands that a
sysadmin is more likely to use that abstract away the calls like
'discover'.