fix: request errors not being checked to see if nil
Some commands were not checking if the error or errors returned by
functions making requests was/were nil before throwing an error message
and exiting as error. This would cause the command to complete
successfully but throw an error as if it did not.

Add checks for err != nil before logging error messages and exiting.