feat(discover): add --overwrite flag
The --overwrite flag to the discover command allows the discover command
to overwrite any existing data that may be present (versus erring
because the data exists, i.e. returning an HTTP 409). This is convenient
when the user knows that the data should be what is passed in the
discover payload, regardless of existing data.