fix: create file if nonexistent; move arg check and config/log init to PersistentPreRunE
In the config set/cluster set commands, properly create the
user-specified config file if nonexistent.

Instead of initializing config via cobra.OnInitialize, move
initialization to the PersistentPreRunE function of the command. The
main reason for this is so that some commands (e.g. config show) don't
ask the user to create nonexistent config files and err instead, wnile
allowing others (e.g. config set) to do ask.

The root command's PersistentPreRunE function, which is inherited by
each child command unless overridden, asks the user to create the config
file if the path the user specified does not exist. If a command needs
to modify this behavior (e.g. the unset commands, which should err if
the config file doesn't exist), that command defines its own
PersistentPreRunE function to err if the path doesn't exist.