ci: use GitHub release notes and release as draft
GitHub release notes seem to be more flexible and extensive than
Goreleaser's, and includes additional features such as listing new
contributors. Instead of the raw commit message "Merge pull request #
from <branch>" being listed as a changelog item, the title of the PR is
used with the link to the PR and author, which is more descriptive.

Also, push the release as a draft so it can be reviewed before actual
release, adding editions as necessary.