feat!: make data/format flags more consistent
The -f flag was being used to pass the path to a file to read the
payload from while -F was being used both to specify an input payload
format and an output format, depending on the command. This can be a bit
confusing, and having consistent and intuitive flag meanings would be
desired.

Now, instead of:

-f/--payload --- Path to payload file
-F/--format  --- Format of payload file (POST/PUT/PATCH) or format of
                 outpout (GET)

the flags are now:

-d/--data          --- Raw payload data or path to payload file
-f/--format-input  --- Format of input payload data
-F/--format-output --- Format of output data

The -d follows the curl paradigm of prepending the flag argument with @
to denote that it is being passed a file path, which can be - to tell
ochami to read the data from standard input. Importantly, input and
output format are differentiated in the case that a command requires
both.

BREAKING CHANGE: -f/--payload changed to -d/--data, format flags are now
-f/--format-input and -F/--format-output