feat: add --help hints and man page references
When errors are logged for a command, an additional log message is
printed, telling the user to invoke --help on the command to see
detailed help output. This message contains the full command invocation
(without flags and arguments) so the user can copy/paste it.

Additionally, a reference to the relevant manual page has been added to
the 'Long' field of each command. If the user is dissatisfied with the
output of --help, they can reference the manual page for even more
detail.