refactor: make lib.go stdin/stdout functions general for testability
Create the ioStream struct that has an io.Reader (stdin) and two
io.Writers (stderr, stdout) that can be used by any function that needs
to read from stdin or write to stdout/stderr. This is so that these
functions can be unit tested by having these set to buffers that can be
written to, read from, and compared.

Functions that have changed are loopYesNo() and askToCreate(), while
prompt() as been removed as it is now not being used anywhere.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>