feat(config): add enable-auth cluster config option
The enable-auth option triggers token checking. It defaults to true, in
which case a token is read and checked locally, erring if invalid or
unset. If set to false for a cluster, the reading/checking of an access
token is disabled for that cluster.

Passing the --no-token flag overrides enable-auth, effectively setting
it to false for the run of the command.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>