test: fix config tests for new dependence on YAML unmarshalling
Since we got rid of calls to koanf's unmarshalling functions in favor of
the the better flexibility of YAML marshalling, we sacrifice the YAML
key checking it offers. However, errors in known keys are still checked,
which is arguably more important than erring on unknown keys.

This commit also makes sure the expected EnableAuth value is true for
YAML cluster configs that do not set it.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>