refactor(discover): move dedup maps outside node loop
Deduplication maps for BMC managers and Systems were declared within the
loop that iterates over each node, which is where duplication checks
happen. This is obviously futile, so the declarations are moved outside
of the loop so that the checks occur properly.

Most testing has used a single unique BMC per node, which is probably
why this was not caught earlier.

Signed-off-by: Devon Bautista <17506592+synackd@users.noreply.github.com>